boletofacilsdk/src/main/java/com/boletobancario/boletofacilsdk/model/entities/CreditCard.java
package com.boletobancario.boletofacilsdk.model.entities;

public class CreditCard extends BaseEntity {
	private String number;
	private String holderName;
	private Integer securityCode;
	private Integer expirationMonth;
	private Integer expirationYear;

	public String getNumber() {
		return number;
	}

	public void setNumber(String number) {
		this.number = number;
	}

	public String getHolderName() {
		return holderName;
	}

	public void setHolderName(String holderName) {
		this.holderName = holderName;
	}

	public Integer getSecurityCode() {
		return securityCode;
	}

	public void setSecurityCode(Integer securityCode) {
		this.securityCode = securityCode;
	}

	public Integer getExpirationMonth() {
		return expirationMonth;
	}

	public void setExpirationMonth(Integer expirationMonth) {
		this.expirationMonth = expirationMonth;
	}

	public Integer getExpirationYear() {
		return expirationYear;
	}

	public void setExpirationYear(Integer expirationYear) {
		this.expirationYear = expirationYear;
	}
}
